clang_delta: Only brute force C++ standard for select heuristics
In the ClangHintsPass, only use the brute-forcing of the "best" C++ standard
(in terms of reduction possibilities) to only two heuristics:
"replace-function-def-with-decl" and "remove-unused-function". For other
heuristics, use the default C++ standard implied by libclang.

This essentially restores how C-Vise used to iterate over these heuristics
before they were migrated to hints; this is expected to save a bit of run time
when --clang-delta-std is not used, since we'll need 5 fewer jobs for each
such heuristic.