Statistics accounts for timed-out/stopping/canceled jobs (#452)
Improve statistics computation to account the time spent on the following jobs:
timed out jobs, jobs that hung in the Pebble task queue, jobs that returned
STOP, and jobs that were canceled because C-Vise was committing to a previously
found successful transformation.

This only improves the information reported at the end of C-Vise execution - no
functional changes to the reduction itself.