test_manager: Specify cwd instead of doing chdir (#457)
When launching the interestingness test, run it with the custom cwd instead of
doing chdir in the C-Vise process.

This fixes a race condition, when a worker process could've been launched in the
middle of check_sanity() and inherit a wrong cwd, resulting in all future jobs
in that worker not being able to read the input file.